Fixed crashed code and deletion of MasterData usage
Struct Master Data is removed, now the Master first distributed the countries to the slaves, who store them in their database. In the computation part, the master sends to each slave the data it wants and waits until all the countries distributed are received. The slave sends the name fo the country and the percentage increase, the Master then reorders the list with an Insertion algorithm and prints the result of that day.